refactor: [#1316] move the rest of the Health Check API integration tests to the server pkg
It's now possible becuase test environments have been exposed in their
server packages.
