+++ tests/common/mod.rs
@@ -1,4 +1,3 @@
 pub mod clock;
 pub mod http;
-pub mod logging;
 pub mod udp;
