--- tests/servers/health_check_api/client.rs
@@ -1,5 +1,0 @@
-use reqwest::Response;
-
-pub async fn get(path: &str) -> Response {
-    reqwest::Client::builder().build().unwrap().get(path).send().await.unwrap()
-}
