--- tests/servers/health_check_api/mod.rs
@@ -1,2 +1,0 @@
-pub mod client;
-pub mod contract;
