--- tests/servers/mod.rs
@@ -1,1 +1,0 @@
-mod health_check_api;
